add write_A_matrix() and latticeCoordinateOutput() into scaling_sim class for outputing the A-matrix elements on every sites at given time; add more output info in main(), such as int steps, sim.t/sim.config.dt etc; add hila directorary into Makefile; A.write(ofstream,false,8) was used for A-matrix output;
(csc git username & email are changed to bitbucket one)

+++ matep.hpp
@@ -19,7 +19,7 @@
 #include <vector>
 
 
-using real_t = double; // or float
+using real_t = float; // or float
 
 class Matep {
 public:
